#minor: reorganize tests & some fixes in ci.yml
