+++ README.md
@@ -4,6 +4,9 @@
 [![Coverage Status](https://coveralls.io/repos/github/umbrellio/laravel-heavy-jobs/badge.svg?branch=master)](https://coveralls.io/github/umbrellio/laravel-heavy-jobs?branch=master)
 [![Latest Stable Version](https://poser.pugx.org/umbrellio/laravel-heavy-jobs/v/stable.png)](https://packagist.org/packages/umbrellio/laravel-heavy-jobs)
 [![Total Downloads](https://poser.pugx.org/umbrellio/laravel-heavy-jobs/downloads.png)](https://packagist.org/packages/umbrellio/laravel-heavy-jobs)
+[![Code Intelligence Status](https://scrutinizer-ci.com/g/umbrellio/laravel-heavy-jobs/badges/code-intelligence.svg?b=master)](https://scrutinizer-ci.com/code-intelligence)
+[![Build Status](https://scrutinizer-ci.com/g/umbrellio/laravel-heavy-jobs/badges/build.png?b=master)](https://scrutinizer-ci.com/g/umbrellio/laravel-heavy-jobs/build-status/master)
+[![Scrutinizer Code Quality](https://scrutinizer-ci.com/g/umbrellio/laravel-heavy-jobs/badges/quality-score.png?b=master)](https://scrutinizer-ci.com/g/umbrellio/laravel-heavy-jobs/?branch=master)
 
 Пакет позволяет сохранять большой payload джобки в стороннем хранилище.
