+++ composer.json
@@ -1,7 +1,12 @@
 {
     "name": "umbrellio/laravel-heavy-jobs",
     "description": "Позволяет отдельно хранить огромные данные задачи из очереди.",
-    "keywords": ["umbrellio", "laravel", "extension"],
+    "keywords": [
+        "umbrellio",
+        "laravel",
+        "extension",
+        "redis"
+    ],
     "minimum-stability": "stable",
     "type": "library",
     "license": "MIT",
@@ -35,7 +40,6 @@
         "umbrellio/code-style-php": "^1.0",
         "phpunit/phpunit": "^8.3|^9.0",
         "sensiolabs/security-checker": "^6.0",
-        "symplify/easy-coding-standard": "^7.2|^8.0|^9.0",
         "php-coveralls/php-coveralls": "^2.1",
         "orchestra/testbench": "^6.2"
     },
